adding root path key

+++ src/cpp/main/dmoz/infectedNodeComputer.h
@@ -13,6 +13,7 @@
 #include <tr1/unordered_map>
 #include <set>
 #include "log4cpp/Category.hh"
+#include "root_path_key.h"
 
 class InfectedNodeComputer {
 public:
@@ -21,7 +22,7 @@ public:
   int findInfectedIndex(long);
   void addInfectedNodes(const std::vector<long>& edges);
   void incrementNumLine();
-  std::tr1::unordered_map<int, std::set<long> > infectedNodes;
+  std::tr1::unordered_map<long, std::set<long>> infectedNodes;
 private:
   std::vector<long> crawlMaxes;
   long maxNodeId;
